+++ rectools/models/base.py
@@ -19,6 +19,7 @@ import typing as tp
 import numpy as np
 import pandas as pd
 import typing_extensions as tpe
+from pydantic import PlainSerializer
 from pydantic_core import PydanticSerializationError
 
 from rectools import AnyIds, Columns, InternalIds
@@ -38,6 +39,20 @@ SemiInternalRecoTriplet = tp.Tuple[AnyIds, InternalIds, Scores]
 RecoTriplet = tp.Tuple[AnyIds, AnyIds, Scores]
 
 RecoTriplet_T = tp.TypeVar("RecoTriplet_T", InternalRecoTriplet, SemiInternalRecoTriplet, RecoTriplet)
+
+
+def _serialize_random_state(rs: tp.Optional[tp.Union[None, int, np.random.RandomState]]) -> tp.Union[None, int]:
+    if rs is None or isinstance(rs, int):
+        return rs
+
+    # NOBUG: We can add serialization using get/set_state, but it's not human readable
+    raise TypeError("`random_state` must be ``None`` or have ``int`` type to convert it to simple type")
+
+
+RandomState = tpe.Annotated[
+    tp.Union[None, int, np.random.RandomState],
+    PlainSerializer(func=_serialize_random_state, when_used="json"),
+]
 
 
 class ModelConfig(BaseConfig):